refactor: move mocks and vitest setup to the root level
