+++ tsconfig.json
@@ -17,6 +17,6 @@
     "outDir": "build",
     "noEmit": true
   },
-  "include": ["src/**/*.ts", "*.config.*"],
+  "include": ["src", "mocks", "*.config.*", "*.setup.*"],
   "exclude": ["node_modules", "build"]
 }
