+++ vitest.setup.ts
@@ -1,6 +1,6 @@
 import { afterAll, afterEach, beforeAll } from 'vitest';
 import { server } from './mocks/server';
-import { logger } from './utils/logger';
+import { logger } from './src/utils/logger';
 
 beforeAll(() => {
   server.listen();
